+++ impress.js
@@ -155,12 +155,13 @@ const loadApplication = async (root) => {
 };
 
 const loadApplications = async () => {
-  const list = await fsp
+  const applications = await fsp
     .readFile('.applications', 'utf8')
     .then((data) => data.split(/[\r\n\s]+/).filter((s) => s.length !== 0))
     .catch(() => [PATH]);
-  for (const path of list) {
-    await loadApplication(path);
+  for (const location of applications) {
+    const dir = path.join(PATH, location);
+    await loadApplication(dir);
   }
 };